rename navbarHeight var and banners module
